refactor(core): Plan B — extract shared Webflow client to tools/core/webflow
- tools/core/webflow/http.py = auth (resolve_token) + the retry/backoff request loop
  (429/5xx + timeout/URLError, "Network error" message) + WebflowApiError
- tools/core/webflow/cms.py = CmsClient (generic reads, find_field, list_items,
  patch_fields [the reusable staged PATCH primitive], ensure_field) + WriteResult/
  CmsItem/CollectionField. The copywriter (Plan C) + future tools build on this.
- tools/summary/webflow_client.py = thin WebflowClient(CmsClient) keeping the
  Summary-field writers (update_item_summary[_parts|_body], ensure_summary_field,
  find_summary_field) as wrappers + re-exporting the dataclasses (single source).
- .importlinter: core.webflow added to the core-is-a-leaf contract.

Hotspot handled: the retry tests patch GLOBAL urllib/time (via wc.urllib/wc.time);
the adapter keeps `import urllib.request`/`import time` so those targets resolve, and
core's request loop uses the same globals + preserves the "Network error" message +
WriteResult shape. cli.py + webflow_designer unchanged.

DEVIATIONS: CmsClient.__init__ is positional-or-keyword dry_run (matches the old
WebflowClient signature; plan said keyword-only). create_item/bulk_create/publish_items
NOT added (no current caller — Script Creation Gate).

Gate: lint-imports 1 kept/0 broken; core.webflow imports standalone; adapter identity +
dry-run verified; webflow/designer/e2e/cli 64 passed (incl. retry tests); full suite
425 passed / 2 pre-existing fails.
